requirements.txt
# Core Framework
streamlit>=1.32.0

# Google Gemini AI
google-generativeai>=0.4.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.27
alembic>=1.13.1

# Vector Database
chromadb>=0.4.22

# Data Processing (using compatible versions)
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.2
xlrd>=2.0.1
pandasql>=0.7.3
scipy>=1.11.0

# Data Visualization
plotly>=5.19.0
matplotlib>=3.8.0
seaborn>=0.13.0

# AI & NLP
langchain>=0.1.11
langchain-google-genai>=0.0.9
sentence-transformers>=2.5.0

# Report Generation
reportlab>=4.1.0
fpdf2>=2.7.0
python-docx>=1.1.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.9.0
pytz>=2024.1
pydantic>=2.6.0
pydantic-settings>=2.2.0

# File Handling
chardet>=5.2.0

# API & HTTP
requests>=2.31.0

# Logging & Monitoring
loguru>=0.7.0

# Type Hints
typing-extensions>=4.9.0

